seperate out to utils and class to manage docker connection
