+++ kit/command_docker_build.py
@@ -1,29 +1,18 @@
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-import docker
-from docker.errors import DockerException
-
-import tarfile
-import json
-from pprint import pprint
 from re import search
 from sys import stderr
 from getpass import getuser
 from os import geteuid, getcwd, getuid, getgid, environ, chdir as change_directory_to
-from platform import system as os_name
-from pathlib import Path
 from dataclasses import dataclass
-from shutil import copyfile
-from typing import Dict, Iterable, Optional
+from pathlib import Path
+from shutil import copyfile, rmtree
+from platform import system as os_name
+from typing import Dict, Iterable
 
-
-class DockerBuildError(Exception):
-    """Exception for something wrong with the docker build."""
-
-    def __init__(self, message, error):
-        super().__init__(message)
-        self.error = error
+from archive import archive_and_compress
+from docker_tools import DockerTools, DockerException
 
 
 @dataclass(frozen=True, init=False)
@@ -38,6 +27,25 @@ def copyfiles(files: Iterable[str], src_dir: str, dst_dir: str) -> None:
     src_dir, dst_dir = Path(src_dir), Path(dst_dir)
     for filename in files:
         copyfile(src_dir / filename, dst_dir / filename)
+
+
+def create_buildargs(environment: Dict[str, str], ID: int) -> Dict[str, str]:
+    if ID:
+        USERID, GROUPID = ID, ID
+    elif os_name() == "Darwin":
+        # Check for Mac OS
+        print("WARNING: DETECTED MAC OS ... ")
+        USERID, GROUPID = 1000, 1000
+    else:
+        USERID, GROUPID = getuid(), getgid()
+
+    # TODO could buildargs and environment be single obj?
+    return {
+        **environment,
+        "UID": str(USERID),
+        "GID": str(GROUPID),
+        "UNAME": environment["USER"],
+    }
 
 
 def print_preamble() -> None:
@@ -64,20 +72,6 @@ http_proxy and https_proxy are set.
         exit(1)
 
 
-def archive_and_compress(
-    name: str, filepaths: Iterable[str], root: Optional[str] = None
-) -> None:
-    """Archive and compress files and directories into tar.gz file"""
-    root = Path(root if root else ".")
-    try:
-        with tarfile.open(name, "x:gz") as tar:
-            for filepath in filepaths:
-                tar.add(root / filepath, arcname=filepath)
-    except FileExistsError as file_exists_error:
-        print(f"Error: The file '{root / filepath}' already exists")
-        # then continue
-
-
 def filter_file_list(file_list: Iterable[str]) -> Iterable[str]:
     """"""
     for filename in file_list:
@@ -88,6 +82,17 @@ def filter_file_list(file_list: Iterable[str]) -> Iterable[str]:
 
 def setup_docker(args):
     """Build the docker for the toolkit"""
+
+    ROOT = args.hekit_root_dir
+
+    # set_stagging_area
+    stagging_path = Path(ROOT) / "__stagging__"
+    stagging_path.mkdir(exist_ok=True)
+
+    if args.clean:
+        rmtree(stagging_path)
+        print("Stagging area deleted")
+        exit(0)
 
     if args.y:
         print_preamble()
@@ -104,26 +109,24 @@ def setup_docker(args):
     if args.check_only:
         try:
             print("CHECKING DOCKER FUNCTIONALITY ...")
-            client = docker.from_env()
+            docker_tools = DockerTools()
         except DockerException as docker_exception:
             print("Docker Error\n", docker_exception, file=stderr)
             exit(1)
 
         print("CHECKING IN-DOCKER CONNECTIVITY ...")
-        docker_checks(client, environment)
+        docker_tools.checks(environment)
         exit(0)
-
-    ROOT = args.hekit_root_dir
-
-    # set_stagging_area
-    stagging_path = Path(ROOT) / "__stagging__"
-    stagging_path.mkdir(exist_ok=True)
 
     parts_tar_gz = stagging_path / "parts.tar.gz"
     if not parts_tar_gz.exists():
         print("MAKING PARTS.TAR.GZ ...")
         with open(ROOT / "docker/which_files.txt") as f:
-            archive_and_compress(parts_tar_gz, filter_file_list(f), root=ROOT)
+            try:
+                archive_and_compress(parts_tar_gz, filter_file_list(f), root=ROOT)
+            except FileExistsError as file_exists_error:
+                print(f"Error: The file '{root / filepath}' already exists")
+                # then continue
 
     copyfiles(
         ("Dockerfile.base", "Dockerfile.toolkit"),
@@ -135,7 +138,7 @@ def setup_docker(args):
 
     try:
         print("CHECKING DOCKER FUNCTIONALITY ...")
-        client = docker.from_env()
+        docker_tools = DockerTools()
     except DockerException as docker_exception:
         print("Docker Error\n", docker_exception, file=stderr)
         exit(1)
@@ -147,21 +150,17 @@ def setup_docker(args):
         f"WARNING: Setting UID/GID of docker user to '{buildargs['UID']}/{buildargs['GID']}'"
     )
 
-    if not image_exists(client, constants.base_label):
+    if not docker_tools.image_exists(constants.base_label):
         print("BUILDING BASE DOCKERFILE ...")
-        response = build_image(
-            client,
-            dockerfile="Dockerfile.base",
-            tag=constants.base_label,
-            buildargs=buildargs,
+        response = docker_tools.build_image(
+            dockerfile="Dockerfile.base", tag=constants.base_label, buildargs=buildargs
         )
         for out in response:
-            print(out, file=outstream)
+            print(out)
 
-    if not image_exists(client, constants.derived_label):
-        print("BUILDING TOOLKIT DOCKERFILE...")
-        response = build_image(
-            client,
+    if not docker_tools.image_exists(constants.derived_label):
+        print("BUILDING TOOLKIT DOCKERFILE ...")
+        response = docker_tools.build_image(
             dockerfile="Dockerfile.toolkit",
             tag=constants.derived_label,
             buildargs=buildargs,
@@ -169,97 +168,7 @@ def setup_docker(args):
         for out in response:
             print(out)
 
-    print("RUN DOCKER CONTAINER...")
+    print("RUN DOCKER CONTAINER ...")
     # Python cannot relinquish control therefore advise
     print("Run container with")
     print(f"docker run -it {constants.derived_label}")
-
-
-def run_script_in_container(
-    containers, environment, scriptpath, image="ubuntu:20.04"
-) -> int:
-    scriptsrc = Path(scriptpath).expanduser().resolve()
-    print(f"ssrc and sdst {scriptsrc}, {scriptdst}")
-    return containers.run(
-        volumes=[f"{scriptsrc}:/script"],
-        detach=True,
-        stream=True,
-        environment=environment,
-        image=image,
-        command="/bin/bash /script",
-    )
-
-
-def docker_checks(client, environment) -> None:
-
-    check_conn = run_script_in_container(
-        # Assume we are in he-toolkit directory
-        client.containers,
-        environment,
-        "./docker/basic-docker-test.sh",
-    )
-    for stream in check_conn.logs(stream=True):
-        print("[CONTAINER]", stream)
-    status_code = check_conn.wait()["StatusCode"]
-    if status_code != 0:
-        print(
-            "In-docker connectivity failing.",
-            f"Return code was '{status_code}'",
-            file=stderr,
-        )
-        exit(1)
-
-
-def image_exists(client, image_name: str) -> bool:
-    """Returns True if image already exists otherwise False"""
-    images = client.images.list(name=image_name)
-    return len(images) > 0
-
-
-def check_build(image):
-    """For use as a decorator.
-    Forward the build info, but check for error in build.
-    If found raise exception"""
-
-    def inner(*args, **kwargs):
-        responses = image(*args, **kwargs)
-        for response in map(json.loads, responses):
-            if "stream" in response.keys():
-                yield response["stream"]
-            elif "error" in response.keys():
-                raise DockerBuildError("Docker build failed", response)
-            else:
-                raise DockerBuildError("Unrecognised stream property", response)
-
-    return inner
-
-
-@check_build
-def build_image(client, dockerfile: str, tag: str, buildargs):
-    """Build images as we like them built"""
-    return client.api.build(
-        dockerfile=dockerfile,
-        rm=True,  # remove intermediates
-        buildargs=buildargs,
-        tag=tag,
-        path=getcwd(),
-    )
-
-
-def create_buildargs(environment: Dict[str, str], ID: int) -> Dict[str, str]:
-    if ID:
-        USERID, GROUPID = ID, ID
-    elif os_name() == "Darwin":
-        # Check for Mac OS
-        print("WARNING: DETECTED MAC OS ... ")
-        USERID, GROUPID = 1000, 1000
-    else:
-        USERID, GROUPID = getuid(), getgid()
-
-    # TODO could buildargs and environment be single obj?
-    return {
-        **environment,
-        "UID": str(USERID),
-        "GID": str(GROUPID),
-        "UNAME": environment["USER"],
-    }
